ptrac: fixed parsing issues

+++ tests/test_pymcnp/test_ptrac/test_history/test_event/test_J_3.py
@@ -10,7 +10,7 @@ class Test_J_3:
             {
                 'next_type': consts.ast.ptrac.history.event.j.EVENT_TYPE,
                 'node': consts.ast.types.INTEGER,
-                'nsx_nsf_nter': consts.ast.types.INTEGER,
+                'nsx_nsf_nter': consts.ast.types.REAL,
                 'ntyn_mtp_angle_branch': consts.ast.types.INTEGER,
                 'ipt': consts.ast.types.INTEGER,
                 'ncl': consts.ast.types.INTEGER,
@@ -21,7 +21,7 @@ class Test_J_3:
             {
                 'next_type': None,
                 'node': consts.ast.types.INTEGER,
-                'nsx_nsf_nter': consts.ast.types.INTEGER,
+                'nsx_nsf_nter': consts.ast.types.REAL,
                 'ntyn_mtp_angle_branch': consts.ast.types.INTEGER,
                 'ipt': consts.ast.types.INTEGER,
                 'ncl': consts.ast.types.INTEGER,
@@ -30,7 +30,7 @@ class Test_J_3:
             {
                 'next_type': consts.ast.ptrac.history.event.j.EVENT_TYPE,
                 'node': None,
-                'nsx_nsf_nter': consts.ast.types.INTEGER,
+                'nsx_nsf_nter': consts.ast.types.REAL,
                 'ntyn_mtp_angle_branch': consts.ast.types.INTEGER,
                 'ipt': consts.ast.types.INTEGER,
                 'ncl': consts.ast.types.INTEGER,
@@ -48,7 +48,7 @@ class Test_J_3:
             {
                 'next_type': consts.ast.ptrac.history.event.j.EVENT_TYPE,
                 'node': consts.ast.types.INTEGER,
-                'nsx_nsf_nter': consts.ast.types.INTEGER,
+                'nsx_nsf_nter': consts.ast.types.REAL,
                 'ntyn_mtp_angle_branch': None,
                 'ipt': consts.ast.types.INTEGER,
                 'ncl': consts.ast.types.INTEGER,
@@ -57,7 +57,7 @@ class Test_J_3:
             {
                 'next_type': consts.ast.ptrac.history.event.j.EVENT_TYPE,
                 'node': consts.ast.types.INTEGER,
-                'nsx_nsf_nter': consts.ast.types.INTEGER,
+                'nsx_nsf_nter': consts.ast.types.REAL,
                 'ntyn_mtp_angle_branch': consts.ast.types.INTEGER,
                 'ipt': None,
                 'ncl': consts.ast.types.INTEGER,
@@ -66,7 +66,7 @@ class Test_J_3:
             {
                 'next_type': consts.ast.ptrac.history.event.j.EVENT_TYPE,
                 'node': consts.ast.types.INTEGER,
-                'nsx_nsf_nter': consts.ast.types.INTEGER,
+                'nsx_nsf_nter': consts.ast.types.REAL,
                 'ntyn_mtp_angle_branch': consts.ast.types.INTEGER,
                 'ipt': consts.ast.types.INTEGER,
                 'ncl': None,
@@ -75,7 +75,7 @@ class Test_J_3:
             {
                 'next_type': consts.ast.ptrac.history.event.j.EVENT_TYPE,
                 'node': consts.ast.types.INTEGER,
-                'nsx_nsf_nter': consts.ast.types.INTEGER,
+                'nsx_nsf_nter': consts.ast.types.REAL,
                 'ntyn_mtp_angle_branch': consts.ast.types.INTEGER,
                 'ipt': consts.ast.types.INTEGER,
                 'ncl': consts.ast.types.INTEGER,
